requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
mysql-connector-python==8.2.0
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Pillow==11.2.1
email-validator==1.1.3 